+++ src/seekstorm_server/server.rs
@@ -1,14 +1,16 @@
 use base64::{engine::general_purpose, Engine};
 use crossbeam_channel::{bounded, select, Receiver};
-use seekstorm::index::{SimilarityType, TokenizerType};
+use seekstorm::{
+    index::{SimilarityType, TokenizerType},
+    ingest::ingest_json,
+};
 use std::{collections::HashMap, env::current_exe, fs, io, path::Path, sync::Arc};
 use tokio::sync::RwLock;
 
 use crate::{
     api_endpoints::{create_apikey_api, create_index_api, delete_apikey_api, open_all_apikeys},
     http_server::{calculate_hash, http_server},
-    ingest::ingest_ndjson,
-    multi_tenancy::{ApikeyObject, ApikeyQuotaObject},
+    multi_tenancy::{get_apikey_hash, ApikeyObject, ApikeyQuotaObject},
 };
 
 fn ctrl_channel() -> Result<Receiver<()>, ctrlc::Error> {
@@ -25,9 +27,9 @@ async fn commandline(sender: crossbeam_channel::Sender<String>) {
         let mut input = String::new();
         match io::stdin().read_line(&mut input) {
             Ok(_n) => {
-                let input_string = input.trim().to_lowercase().to_owned();
+                let input_string = input.trim().to_owned();
                 let _ = sender.send(input_string.to_string());
-                if input_string == "quit" {
+                if input_string.to_lowercase() == "quit" {
                     break;
                 }
             }
@@ -119,49 +121,98 @@ pub(crate) async fn initialize(params: HashMap<String, String>) {
 
                 let m=message.unwrap();
                 let parameter:Vec<&str>=m.split_whitespace().collect();
-                let command=if parameter.is_empty() {""} else {parameter[0]};
+                let command=if parameter.is_empty() {""} else {&parameter[0].to_lowercase()};
 
                 match command
                 {
                     "ingest" =>
                     {
+                        if parameter.len()==1 || parameter.len()==4 {
+                            let apikey_list_clone2=apikey_list_clone.clone();
+                            let apikey_option=if parameter.len()>2 {
+                                get_apikey_hash(parameter[2].to_string(), &apikey_list_clone2).await
+                            } else {
+                                None
+                            };
 
-                        let apikey_quota_object=ApikeyQuotaObject {..Default::default()};
+                            let mut apikey_list_mut = apikey_list_clone.write().await;
+                            let apikey_object_option=if parameter.len()>2 {
 
-                        let mut apikey_list_mut = apikey_list_clone.write().await;
-                        let apikey_object=create_apikey_api(
-                            &index_path,
-                            apikey_quota_object,
-                            &demo_api_key,
-                            &mut apikey_list_mut,
-                        );
+                                if let Some(apikey_hash) = apikey_option
+                                {
+                                    apikey_list_mut.get_mut(&apikey_hash)
+                                } else if parameter[2]==demo_api_key_base64{
+                                    let apikey_quota_object=ApikeyQuotaObject {..Default::default()};
+                                    Some(create_apikey_api(
+                                        &index_path,
+                                        apikey_quota_object,
+                                        &demo_api_key,
+                                        &mut apikey_list_mut,
+                                    ))
+                                } else{
+                                    None
+                                }
+                            } else {
+                                let apikey_quota_object=ApikeyQuotaObject {..Default::default()};
+                                Some(create_apikey_api(
+                                    &index_path,
+                                    apikey_quota_object,
+                                    &demo_api_key,
+                                    &mut apikey_list_mut,
+                                ))
+                            };
 
-                        let wikipedia_schema_json = r#"
-                        [{"field":"title","field_type":"Text","stored":true,"indexed":true,"boost":10.0},
-                        {"field":"body","field_type":"Text","stored":true,"indexed":true},
-                        {"field":"url","field_type":"Text","stored":true,"indexed":false}]"#;
-                        let schema = serde_json::from_str(wikipedia_schema_json).unwrap();
+                            if let Some(apikey_object) = apikey_object_option {
 
-                        let index_id = create_index_api(
-                            &index_path,
-                            "wikipedia".to_string(),
-                            schema,
-                            SimilarityType::Bm25fProximity,
-                            TokenizerType::UnicodeAlphanumeric,
-                            apikey_object,
-                        );
+                                let index_id=if parameter.len()>3 {
+                                    parameter[3].parse().unwrap_or(0)
+                                } else  if apikey_object.index_list.is_empty() || !apikey_object.index_list.contains_key(&0) {
+                                        let wikipedia_schema_json = r#"
+                                        [{"field":"title","field_type":"Text","stored":true,"indexed":true,"boost":10.0},
+                                        {"field":"body","field_type":"Text","stored":true,"indexed":true},
+                                        {"field":"url","field_type":"Text","stored":true,"indexed":false}]"#;
+                                        create_index_api(
+                                            &index_path,
+                                            "demo_index".to_string(),
+                                            serde_json::from_str(wikipedia_schema_json).unwrap(),
+                                            SimilarityType::Bm25fProximity,
+                                            TokenizerType::UnicodeAlphanumeric,
+                                            apikey_object,
+                                        )
+                                    } else {
+                                        0
+                                    };
 
-                        let apikey_hash = calculate_hash(&demo_api_key) as u128;
-                        let apikey_object=apikey_list_mut.get(&apikey_hash).unwrap();
-                        let index_arc=apikey_object.index_list.get(&index_id).unwrap();
-                        let index_arc_clone=index_arc.clone();
+                                if let Some(index_arc) = apikey_object.index_list.get(&index_id) {
+                                    let index_arc_clone=index_arc.clone();
 
-                        drop(apikey_list_mut);
+                                    drop(apikey_list_mut);
 
-                        let data_path=if parameter.len()>1 {parameter[1]} else {"wiki-articles.json"};
-                        ingest_ndjson(index_arc_clone,data_path).await;
+                                    let data_path_str=if parameter.len()>1 {
+                                        parameter[1]
+                                    } else {
+                                        "wiki-articles.json"
+                                    };
+                                    let mut data_path = Path::new(&data_path_str);
+                                    let mut absolute_path = current_exe().unwrap();
+                                    if !data_path.is_absolute() {
+                                        absolute_path.pop();
+                                        absolute_path.push(data_path_str);
+                                        data_path = &absolute_path;
+                                    }
 
-                        println!("Set the 'individual API key' in test_api.rest to '{}' when testing the REST API endpoints",demo_api_key_base64);
+                                    ingest_json(index_arc_clone,data_path).await;
+
+                                    println!("Set the 'individual API key' in test_api.rest to '{}' when testing the REST API endpoints",if parameter.len()>2 {parameter[2]}else{&demo_api_key_base64});
+                                } else {
+                                    println!("Index not found {}",index_id);
+                                }
+                            } else {
+                                println!("API key not found {}. Create a valid API key first via REST API or use the demo API key {}",parameter[2],demo_api_key_base64);
+                            }
+                        } else {
+                            println!("Missing parameters: ingest [data_path] [apikey] [index_id]");
+                        }
                     },
 
                     "delete" =>